Revert "Draper changes (#13)"
This reverts commit cc42a65e6e0cdc3be88150553cd7c4108868c14c.

+++ Radamsa/vmf/src/modules/common/mutator/RadamsaByteMutatorBase.hpp
@@ -61,7 +61,7 @@ public:
         constexpr size_t MINIMUM_UPPER_LIMIT{0x2u};
         constexpr size_t MAXIMUM_UPPER_LIMIT{0x20000u};
 
-        size_t randomStop{rand->randBetween(0ul, static_cast<unsigned long>(MINIMUM_UPPER_LIMIT))};
+        size_t randomStop{rand->randBetween(0u, MINIMUM_UPPER_LIMIT)};
         size_t randomUpperLimit{MINIMUM_UPPER_LIMIT};
 
         while(randomStop != 0u)
@@ -70,10 +70,10 @@ public:
                 break;
 
             randomUpperLimit <<= 1u;
-            randomStop = rand->randBetween(0ul, static_cast<unsigned long>(MINIMUM_UPPER_LIMIT));
+            randomStop = rand->randBetween(0u, MINIMUM_UPPER_LIMIT);
         }
 
-        return rand->randBetween(0ul, static_cast<unsigned long>(randomUpperLimit)) + 1u; // We add one to the return value in order to account for the case where the random upper value is zero.
+        return rand->randBetween(0u, randomUpperLimit) + 1u; // We add one to the return value in order to account for the case where the random upper value is zero.
     }
 
     vector<uint8_t> encodeUtf8(char32_t cp) {
