+++ Radamsa/vmf/src/modules/common/mutator/RadamsaIncrementByteMutator.cpp
@@ -142,15 +142,15 @@ void RadamsaIncrementByteMutator::mutateTestCase(StorageModule& storage, Storage
 
     // Select a random byte to circularly increment.
 
-    const unsigned long lower{0ul};
+    const size_t lower{0u};
     const size_t upper{originalSize - 1u};
-    const unsigned long maximumRandomIndexValue{static_cast<unsigned long>(originalSize - minimumSeedIndex)};
+    const size_t maximumRandomIndexValue{originalSize - minimumSeedIndex};
     const size_t randomIndexToIncrement{
                                     std::clamp(
-                                        static_cast<size_t>(rand->randBetween(
+                                        rand->randBetween(
                                             lower,
-                                            maximumRandomIndexValue)) + minimumSeedIndex,
-                                        static_cast<size_t>(lower),
+                                            maximumRandomIndexValue) + minimumSeedIndex,
+                                        lower,
                                         upper
                                     )
     };
