+++ Radamsa/vmf/src/modules/common/mutator/RadamsaSwapLineMutator.cpp
@@ -145,12 +145,12 @@ void RadamsaSwapLineMutator::mutateTestCase(StorageModule& storage, StorageEntry
 
     // Select random line to copy from/to.
 
-    const unsigned long minimumRandomLineOffset{0ul};
+    const size_t minimumRandomLineOffset{0u};
 
     const size_t firstRandomLineIndex{
-                                static_cast<size_t>(rand->randBetween(
+                                rand->randBetween(
                                                     minimumRandomLineOffset,
-                                                    static_cast<unsigned long>(numberOfLinesAfterIndex - 1u)))};
+                                                    numberOfLinesAfterIndex - 1u)};
 
     const size_t totalNumberOfLines{
                                 GetNumberOfLinesAfterIndex(
